+++ babeld.man
@@ -309,14 +309,6 @@ This specifies the name of the file to which
 .B babeld
 writes out its process id, and is equivalent to the command-line option
 .BR \-I .
-.TP
-.BI first-table-number " table"
-This specifies the index of the first routing table to use for
-source-specific routes.  The default is 10.
-.TP
-.BI first-rule-priority " priority"
-This specifies smallest (highest) rule priority used with source-specific
-routes.  The default is 100.
 .SS Interface configuration
 An interface is configured by a line with the following format:
 .IP
