+++ kernel_netlink.c
@@ -934,14 +934,6 @@ kernel_interface_channel(const char *ifname, int ifindex)
         return -1;
 }
 
-/* Return true if we cannot handle disambiguation ourselves. */
-
-int
-kernel_disambiguate(int v4)
-{
-    return !v4 && has_ipv6_subtrees;
-}
-
 int
 kernel_has_ipv6_subtrees(void)
 {
@@ -1024,7 +1016,7 @@ kernel_route(int operation, int table,
 
 
     ipv4 = v4mapped(gate);
-    use_src = (!is_default(src, src_plen) && kernel_disambiguate(ipv4));
+    use_src = !is_default(src, src_plen);
 
     kdebugf("kernel_route: %s %s from %s "
             "table %d metric %d dev %d nexthop %s\n",
@@ -1448,64 +1440,6 @@ filter_addresses(struct nlmsghdr *nh, struct kernel_addr *addr)
 }
 
 static int
-filter_kernel_rules(struct nlmsghdr *nh, struct kernel_rule *rule)
-{
-    int len, has_priority = 0, has_table = 0;
-    struct rtmsg *rtm = NULL;
-    struct rtattr *rta = NULL;
-    int is_v4 = 0;
-
-    len = nh->nlmsg_len;
-
-    rtm = (struct rtmsg*)NLMSG_DATA(nh);
-    len -= NLMSG_LENGTH(0);
-
-    rule->src_plen = rtm->rtm_src_len;
-    memset(rule->src, 0, sizeof(rule->src));
-    rule->table = rtm->rtm_table;
-
-    if(rtm->rtm_family != AF_INET && rtm->rtm_family != AF_INET6) {
-        kdebugf("filter_rules: Unknown family: %d\n", rtm->rtm_family);
-        return -1;
-    }
-    is_v4 = rtm->rtm_family == AF_INET;
-
-    rta = RTM_RTA(rtm);
-    len -= NLMSG_ALIGN(sizeof(*rtm));
-
-    for(; RTA_OK(rta, len); rta = RTA_NEXT(rta, len)) {
-        switch(rta->rta_type) {
-        case FRA_UNSPEC: break;
-        case FRA_SRC:
-            rule->src_plen = GET_PLEN(rtm->rtm_src_len, is_v4);
-            COPY_ADDR(rule->src, rta, is_v4);
-            break;
-        case FRA_PRIORITY:
-            rule->priority = *(unsigned int*)RTA_DATA(rta);
-            has_priority = 1;
-            break;
-        case FRA_TABLE:
-            rule->table = *(int*)RTA_DATA(rta);
-            has_table = 1;
-            break;
-        default:
-            kdebugf("filter_rules: Unknown rule attribute: %d.\n",
-                    rta->rta_type);
-            break;
-        }
-    }
-
-    kdebugf("filter_rules: from %s prio %d table %d\n",
-            format_prefix(rule->src, rule->src_plen),
-            has_priority ? rule->priority : -1, rule->table);
-
-    if(!has_priority || !has_table)
-        return 0;
-
-    return 1;
-}
-
-static int
 filter_netlink(struct nlmsghdr *nh, struct kernel_filter *filter)
 {
     int rc;
@@ -1513,7 +1447,6 @@ filter_netlink(struct nlmsghdr *nh, struct kernel_filter *filter)
         struct kernel_route route;
         struct kernel_addr addr;
         struct kernel_link link;
-        struct kernel_rule rule;
     } u;
 
     switch(nh->nlmsg_type) {
@@ -1535,12 +1468,6 @@ filter_netlink(struct nlmsghdr *nh, struct kernel_filter *filter)
         rc = filter_addresses(nh, &u.addr);
         if(rc <= 0) break;
         return filter->addr(&u.addr, filter->addr_closure);
-    case RTM_NEWRULE:
-    case RTM_DELRULE:
-        if(!filter->rule) break;
-        rc = filter_kernel_rules(nh, &u.rule);
-        if(rc <= 0) break;
-        return filter->rule(&u.rule, filter->rule_closure);
     default:
         kdebugf("filter_netlink: unexpected message type %d\n",
                 nh->nlmsg_type);
@@ -1569,143 +1496,4 @@ kernel_callback(struct kernel_filter *filter)
         kernel_setup_socket(1);
 
     return 0;
-}
-
-
-/* Routing table's rules */
-
-int
-add_rule(int prio, const unsigned char *src_prefix, int src_plen, int table)
-{
-    char buffer[64] = {0}; /* 56 needed */
-    struct nlmsghdr *message_header = (void*)buffer;
-    struct rtmsg *message = NULL;
-    struct rtattr *current_attribute = NULL;
-    int is_v4 = v4mapped(src_prefix);
-    int addr_size = is_v4 ? sizeof(struct in_addr) : sizeof(struct in6_addr);
-
-    kdebugf("Add rule v%c prio %d from %s\n", is_v4 ? '4' : '6', prio,
-            format_prefix(src_prefix, src_plen));
-
-    if(is_v4) {
-        src_prefix += 12;
-        src_plen -= 96;
-        if(src_plen < 0) {
-            errno = EINVAL;
-            return -1;
-        }
-    }
-
-#if RTA_ALIGNTO != NLMSG_ALIGNTO
-#error "RTA_ALIGNTO != NLMSG_ALIGNTO"
-#endif
-
-    /* Set the header */
-    message_header->nlmsg_flags = NLM_F_REQUEST | NLM_F_CREATE | NLM_F_EXCL;
-    message_header->nlmsg_type  = RTM_NEWRULE;
-    message_header->nlmsg_len   = NLMSG_ALIGN(sizeof(struct nlmsghdr));
-
-    /* Append the message */
-    message = NLMSG_DATA(message_header);
-    message->rtm_family = is_v4 ? AF_INET : AF_INET6;
-    message->rtm_dst_len = 0;
-    message->rtm_src_len = src_plen;
-    message->rtm_tos = 0;
-    message->rtm_table = table;
-    message->rtm_protocol = RTPROT_BABEL;
-    message->rtm_scope = RT_SCOPE_UNIVERSE;
-    message->rtm_type = RTN_UNICAST;
-    message->rtm_flags = 0;
-    message_header->nlmsg_len += NLMSG_ALIGN(sizeof(struct rtmsg));
-
-    /* Append each attribute */
-    current_attribute = RTM_RTA(message);
-    /* prio */
-    current_attribute->rta_len = RTA_LENGTH(sizeof(int));
-    current_attribute->rta_type = FRA_PRIORITY;
-    *(int*)RTA_DATA(current_attribute) = prio;
-
-    message_header->nlmsg_len += current_attribute->rta_len;
-    current_attribute = (void*)
-        ((char*)current_attribute + current_attribute->rta_len);
-
-    /* src */
-    current_attribute->rta_len = RTA_LENGTH(addr_size);
-    current_attribute->rta_type = FRA_SRC;
-    memcpy(RTA_DATA(current_attribute), src_prefix, addr_size);
-
-    message_header->nlmsg_len += current_attribute->rta_len;
-    current_attribute = (void*)
-        ((char*)current_attribute + current_attribute->rta_len);
-
-    /* send message */
-    if(message_header->nlmsg_len > 64) {
-        errno = EINVAL;
-        return -1;
-    }
-    return netlink_talk(message_header);
-}
-
-int
-flush_rule(int prio, int family)
-{
-    char buffer[64] = {0}; /* 36 needed */
-    struct nlmsghdr *message_header = (void*)buffer;
-    struct rtmsg *message = NULL;
-    struct rtattr *current_attribute = NULL;
-
-    kdebugf("Flush rule v%c prio %d\n", family == AF_INET ? '4' : '6', prio);
-
-#if RTA_ALIGNTO != NLMSG_ALIGNTO
-#error "RTA_ALIGNTO != NLMSG_ALIGNTO"
-#endif
-
-    /* Set the header */
-    message_header->nlmsg_flags = NLM_F_REQUEST;
-    message_header->nlmsg_type  = RTM_DELRULE;
-    message_header->nlmsg_len   = NLMSG_ALIGN(sizeof(struct nlmsghdr));
-
-    /* Append the message */
-    message = NLMSG_DATA(message_header);
-    message->rtm_family = family;
-    message->rtm_dst_len = 0;
-    message->rtm_src_len = 0;
-    message->rtm_tos = 0;
-    message->rtm_table = 0;
-    message->rtm_protocol = RTPROT_BABEL;
-    message->rtm_scope = RT_SCOPE_UNIVERSE;
-    message->rtm_type = RTN_UNSPEC;
-    message->rtm_flags = 0;
-    message_header->nlmsg_len += NLMSG_ALIGN(sizeof(struct rtmsg));
-
-    /* Append each attribute */
-    current_attribute = RTM_RTA(message);
-    /* prio */
-    current_attribute->rta_len = RTA_LENGTH(sizeof(int));
-    current_attribute->rta_type = FRA_PRIORITY;
-    *(int*)RTA_DATA(current_attribute) = prio;
-
-    message_header->nlmsg_len += current_attribute->rta_len;
-    current_attribute = (void*)
-        ((char*)current_attribute + current_attribute->rta_len);
-
-    /* send message */
-    if(message_header->nlmsg_len > 64) {
-        errno = EINVAL;
-        return -1;
-    }
-    return netlink_talk(message_header);
-}
-
-int
-change_rule(int new_prio, int old_prio,
-            const unsigned char *src, int plen, int table)
-{
-    int rc;
-    kdebugf("/Swap: ");
-    rc = add_rule(new_prio, src, plen, table);
-    if(rc < 0)
-        return rc;
-    kdebugf("\\Swap: ");
-    return flush_rule(old_prio, v4mapped(src) ? AF_INET : AF_INET6);
 }
