--- rule.h
@@ -1,33 +1,0 @@
-/*
-Copyright (c) 2015 by Matthieu Boutier and Juliusz Chroboczek.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-*/
-
-#define SRC_TABLE_NUM 10
-
-extern int src_table_idx; /* number of the first table */
-extern int src_table_prio; /* first prio range */
-
-/* Return the number of the table using src_plen, allocate the table in the
-   kernel if necessary. */
-int find_table(const unsigned char *dest, unsigned short plen,
-               const unsigned char *src, unsigned short src_plen);
-void release_tables(void);
-int check_rules(void);
